backend/requirements.txt
fastapi
uvicorn
google-genai
google-auth-oauthlib
google-api-python-client
python-jose[cryptography]
python-dotenv
requests
pydantic
sqlalchemy
